Add a resource for retrieving diffs
The `DiffResource`, available at
`a/projects/<projectName>/commits/<sha>/diff/`, yields the diff of the
commit with the given sha. Due to limitations within the JGit library,
we are only able to compute diffs for commits which have a single
parent.

Testing Done:
Hit the API and saw the diff returned.

Reviewed at https://reviews.reviewboard.org/r/8920/