Add resources for fetching blobs & blob info
This patch adds two new resources: `GetBlobInfo` and `GetBlobContent`.
The former returns whether or not a blob with the given ID exists and
the latter returns the content of a blob with the given ID as base64-
encoded data.

Testing Done:
Retrieved a file via the API.
Verified invalid blob IDs, etc. resulted in a 404.

Reviewed at https://reviews.reviewboard.org/r/8923/